+++ src/components/card/notFound/styles.css
@@ -1,3 +1,3 @@
-img{
-    width: 7rem;
+img {
+  width: 7rem;
 }
