+++ src/context/menuContext/index.js
@@ -1,5 +1,4 @@
 import { createContext } from "react";
 
-const MenuContext = createContext()
-export default MenuContext
-
+const MenuContext = createContext();
+export default MenuContext;
