+++ src/helper/helpSweetAlerts/helpSweetAlertMsg.js
@@ -1,10 +1,10 @@
 import Swal from "sweetalert2";
 
-export function sweetAlertMsg(icon,msg,title) {
+export function helpSweetAlertMsg(icon, msg, title) {
   Swal.fire({
     icon: `${icon}`,
     title: `${title}`,
     text: `${msg}`,
-    timer: 3000
+    timer: 3000,
   });
 }
